[Chore]: S3Uploader 인터페이스 및 구현체 이름 변경.
