+++ src/main/java/com/tenten/linkhub/domain/space/service/SpaceImageUploader.java
@@ -2,7 +2,7 @@ package com.tenten.linkhub.domain.space.service;
 
 import com.tenten.linkhub.global.aws.dto.ImageInfo;
 import com.tenten.linkhub.global.aws.dto.ImageSaveRequest;
-import com.tenten.linkhub.global.aws.s3.S3Uploader;
+import com.tenten.linkhub.global.aws.s3.ImageFileUploader;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
@@ -15,7 +15,7 @@ public class SpaceImageUploader {
     private static final String SPACE_IMAGE_FOLDER = "space-image/";
     private static final String SPACE_DEFAULT_IMAGE_PATH = "https://team-10-bucket.s3.ap-northeast-2.amazonaws.com/space-image/space-default.png";
 
-    private final S3Uploader s3Uploader;
+    private final ImageFileUploader imageFileUploader;
 
     public Optional<ImageInfo> getNewImageInfoOrEmptyImageInfo(MultipartFile file){
         if (file == null){
@@ -23,7 +23,7 @@ public class SpaceImageUploader {
         }
 
         ImageSaveRequest imageSaveRequest = ImageSaveRequest.of(file, SPACE_IMAGE_FOLDER);
-        return Optional.ofNullable(s3Uploader.saveImage(imageSaveRequest));
+        return Optional.ofNullable(imageFileUploader.saveImage(imageSaveRequest));
     }
 
     public ImageInfo getNewImageInfoOrDefaultImageInfo(MultipartFile file) {
@@ -32,7 +32,7 @@ public class SpaceImageUploader {
         }
 
         ImageSaveRequest imageSaveRequest = ImageSaveRequest.of(file, SPACE_IMAGE_FOLDER);
-        return s3Uploader.saveImage(imageSaveRequest);
+        return imageFileUploader.saveImage(imageSaveRequest);
     }
 
 }
